4-scan_ai-Manus/clean_project/src/requirements.txt
attr==0.3.2
ConfigParser==7.2.0
contextlib2==21.6.0
coremltools==8.3.0
cryptography==45.0.4
Cython==3.1.2
dl==0.1.0
docutils==0.21.2
evidently==0.7.7
fastapi==0.115.12
filelock==3.18.0
Flask==3.1.1
flask_cors==6.0.0
flask_jwt_extended==4.7.1
flask_login==0.6.3
flask_sqlalchemy==3.1.1
fpdf==1.7.2
fpdf2==2.8.3
HTMLParser==0.0.2
ipython==8.12.3
ipywidgets==8.1.7
Jinja2==3.1.6
jnius==1.1.0
jose==1.0.0
keyring==25.6.0
langfuse==3.0.1
matplotlib==3.10.3
mock==5.2.0
mysql_connector_repackaged==0.3.1
numpy==2.3.0
onnx==1.18.0
onnx_tf==1.10.0
openpyxl==3.1.5
pandas==2.3.0
passlib==1.7.4
pdfkit==1.0.0
Pillow==11.2.1
playwright==1.52.0
prometheus_client==0.22.1
protobuf==6.31.1
psutil==7.0.0
psycopg2==2.9.10
pydantic==2.11.7
PyJWT==2.10.1
PyJWT==2.10.1
pyOpenSSL==25.1.0
pyotp==2.9.0
pytest==8.3.5
python-dotenv==1.1.0
python_magic==0.4.27
PyYAML==6.0.2
PyYAML==6.0.2
railroad==0.5.0
redis==6.2.0
reportlab==4.4.1
schedule==1.2.2
scikit_learn==1.7.0
seaborn==0.13.2
Sphinx==8.2.3
SQLAlchemy==2.0.41
starlette==0.47.0
tensorflow==2.19.0
tensorflow_datasets==4.9.9
tensorflow_hub==0.16.1
thread==2.0.5
torch==2.7.1
torchvision==0.22.1
tornado==6.5
transformers==4.52.4
trove_classifiers==2025.5.9.12
urllib3_secure_extra==0.1.0
uvicorn==0.34.3
Werkzeug==3.1.3
xmlrpclib==1.0.1
